# swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"responseType": {
				"properties": {
					"message": {
						"type": "string"
					},
					"Status_Code": {
						"type": "number",
						"format": "double"
					},
					"result": {
						"additionalProperties": false,
						"type": "object"
					}
				},
				"required": [
					"message",
					"Status_Code"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Jwt using "
			}
		}
	},
	"info": {
		"title": "typescript-project",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/users": {
			"get": {
				"operationId": "User_detailsfun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		},
		"/admin/user/create": {
			"post": {
				"operationId": "New_Users",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"religion": {
										"type": "string"
									},
									"father_name": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"gender": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"mobileNum": {
										"type": "number",
										"format": "double"
									},
									"email": {
										"type": "string"
									},
									"lname": {
										"type": "string"
									},
									"fname": {
										"type": "string"
									},
									"role": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"religion",
									"father_name",
									"username",
									"gender",
									"password",
									"mobileNum",
									"email",
									"lname",
									"fname",
									"role"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/user/update/{id}": {
			"put": {
				"operationId": "Update_userfun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"religion": {
										"type": "string"
									},
									"father_name": {
										"type": "string"
									},
									"gender": {
										"type": "string"
									},
									"password": {
										"type": "string"
									},
									"mobileNum": {
										"type": "number",
										"format": "double"
									},
									"email": {
										"type": "string"
									},
									"lname": {
										"type": "string"
									},
									"fname": {
										"type": "string"
									},
									"role": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"religion",
									"father_name",
									"gender",
									"password",
									"mobileNum",
									"email",
									"lname",
									"fname",
									"role"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/user/delete/{id}": {
			"delete": {
				"operationId": "Delete_Userfun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/admin/user/login": {
			"post": {
				"operationId": "UserLoginFun",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									},
									"password": {
										"type": "string"
									}
								},
								"required": [
									"email",
									"password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/admin/hello": {
			"get": {
				"operationId": "HelloUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/responseType"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Admin"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": []
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}